Update patch set 7
Patch Set 7:

I should be more clear, I was trying to evaluate the set of files being added here in this directory.  Maybe it's appropriate to ask, who will be clients of this library?  In other words, should it be packaged up and exposed for convenient use by others, or is it just a piece-part of this project and should be kept hidden from other users?

Patch-set: 7