Update patch set 8
Patch Set 8:

(4 comments)

I know I'm repeating myself, but this collection of interfaces and classes looks really strange.  Maybe if we were in the same room with a whiteboard for an hour we could talk it thru and I could tell you why I'm reacting so allergically to what you have submitted here.

This just doesn't look good to me, so I feel unable to give +1 here.

Patch-set: 8
Label: Code-Review=0