Update patch set 9
Patch Set 9:

(4 comments)

Hi Chris,

I would not mind going through these changes. Please check the latest. I think the exception handling is simplified now without extra interface. In addition removed the setters in the interfaces I think that there was a comment earlier about adding setter and it was incorrectly added to the interface. If you are still not satisfied please shoot me an email and we can discuss using zoom.

Patch-set: 9