Update patch set 9
Patch Set 9:

Please start by adding a brief purpose/summary to the top of each interface.  For example I read "CreateRTURequest" as a verb, but to you it's a noun (a request in flight).  Don't change that name for me. A sentence at the top would have removed ambiguity. 

I'm not sure it adds significant value to define interfaces for POJOs, plain old java object, aka bean classes that have some private data with getters and setters.  Consider the CDS domain package which has over 50 classes.  If I had an interface AND an implementation for each one that would really add to the work but not value.

Interfaces are very important if you expect to have alternate implementations.  Using CDS again as an example, I defined the client in terms of an interface so developers can use a real or a mock implementation.  

Interfaces are also very important in Spring contexts, where you can configure what implementation gets autowired.

Thanks for listening.

Patch-set: 9
Reviewer: Gerrit User 4 <4@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0