Update patch set 2
Patch Set 2:

A jar with a class with a main can be built as a special kind of executable jar. None of the Acumos java libraries have been done like this, because all the existing/old libraries are consumed as libraries in the build process, not used by users to launch processes.

Patch-set: 2